requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
hashids==1.3.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
